// ya-ryadom-ui/src/utils/enums/tabs.enum.ts
export enum TABS {
    APPLICATIONS_I_WILL_GO = 'APPLICATIONS_I_WILL_GO',
    APPLICATIONS_REJECTED = 'APPLICATIONS_REJECTED',
    SENT_APPLICATIONS = 'SENT_APPLICATIONS',
    VISITED_APPLICATIONS = 'VISITED_APPLICATIONS',
    WANT_VISIT_APPLICATIONS = 'WANT_VISIT_APPLICATIONS',
    CREATED_APPLICATIONS = 'CREATED_APPLICATIONS',
    EVENTS_MAP = 'EVENTS_MAP',
    EVENTS_LIST = 'EVENTS_LIST',
}